+++ test/TestApp/src/TestApp.jl
@@ -17,4 +17,7 @@ module TestApp
   include("../../AdaptivityUnstructuredTests.jl")
   include("../../AdaptivityMultiFieldTests.jl")
   include("../../BlockSparseMatrixAssemblersTests.jl")
+  include("../../VisualizationTests.jl")
+  include("../../AutodiffTests.jl")
+  include("../../ConstantFESpacesTests.jl")
 end